requirements.txt
# Flask web framework
Flask==3.0.0
Werkzeug==3.0.1

# Google Cloud libraries
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.18.0

# Excel file generation
openpyxl==3.1.2

# Production WSGI server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0

# Utilities
python-dateutil==2.8.2
